requirements.txt
streamlit
pandas
openpyxl  #Primary Database
pytesseract
easyocr
google-generativeai
pillow
numpy
pymupdf
